ZarpadoAPI/backend/models/prenda.py
from typing import Optional
from pydantic import BaseModel

class PrendaCreate(BaseModel):
    nombre: str
    tipo: str
    descripcion: str
    marca: str

class PrendaOut(BaseModel):
    id: str
    nombre: str
    tipo: str
    descripcion: str
    marca: str
    image_path: Optional[str] = None
