ZarpadoAPI/backend/requirements.txt
fastapi
uvicorn
pymongo
neo4j
python-multipart
pillow
email-validator
google-genai
python-dotenv
